+++ src/resources/completions.ts
@@ -293,6 +293,8 @@ export interface CompletionCreateParamsBase {
   seed?: number | null;
 
   /**
+   * Not supported with latest reasoning models `o3` and `o4-mini`.
+   *
    * Up to 4 sequences where the API will stop generating further tokens. The
    * returned text will not contain the stop sequence.
    */