+++ src/resources/shared.ts
@@ -1,6 +1,24 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
+export type AllModels =
+  | (string & {})
+  | ChatModel
+  | 'o1-pro'
+  | 'o1-pro-2025-03-19'
+  | 'computer-use-preview'
+  | 'computer-use-preview-2025-03-11';
+
 export type ChatModel =
+  | 'gpt-4.1'
+  | 'gpt-4.1-mini'
+  | 'gpt-4.1-nano'
+  | 'gpt-4.1-2025-04-14'
+  | 'gpt-4.1-mini-2025-04-14'
+  | 'gpt-4.1-nano-2025-04-14'
+  | 'o4-mini'
+  | 'o4-mini-2025-04-16'
+  | 'o3'
+  | 'o3-2025-04-16'
   | 'o3-mini'
   | 'o3-mini-2025-01-31'
   | 'o1'
@@ -9,11 +27,6 @@ export type ChatModel =
   | 'o1-preview-2024-09-12'
   | 'o1-mini'
   | 'o1-mini-2024-09-12'
-  | 'computer-use-preview'
-  | 'computer-use-preview-2025-02-04'
-  | 'computer-use-preview-2025-03-11'
-  | 'gpt-4.5-preview'
-  | 'gpt-4.5-preview-2025-02-27'
   | 'gpt-4o'
   | 'gpt-4o-2024-11-20'
   | 'gpt-4o-2024-08-06'
@@ -23,6 +36,10 @@ export type ChatModel =
   | 'gpt-4o-audio-preview-2024-12-17'
   | 'gpt-4o-mini-audio-preview'
   | 'gpt-4o-mini-audio-preview-2024-12-17'
+  | 'gpt-4o-search-preview'
+  | 'gpt-4o-mini-search-preview'
+  | 'gpt-4o-search-preview-2025-03-11'
+  | 'gpt-4o-mini-search-preview-2025-03-11'
   | 'chatgpt-4o-latest'
   | 'gpt-4o-mini'
   | 'gpt-4o-mini-2024-07-18'
@@ -171,16 +188,23 @@ export interface Reasoning {
    * supported values are `low`, `medium`, and `high`. Reducing reasoning effort can
    * result in faster responses and fewer tokens used on reasoning in a response.
    */
-  effort: ReasoningEffort | null;
+  effort?: ReasoningEffort | null;
 
   /**
-   * **o-series models only**
+   * @deprecated **Deprecated:** use `summary` instead.
    *
    * A summary of the reasoning performed by the model. This can be useful for
-   * debugging and understanding the model's reasoning process. One of `concise` or
-   * `detailed`.
+   * debugging and understanding the model's reasoning process. One of `auto`,
+   * `concise`, or `detailed`.
    */
-  generate_summary?: 'concise' | 'detailed' | null;
+  generate_summary?: 'auto' | 'concise' | 'detailed' | null;
+
+  /**
+   * A summary of the reasoning performed by the model. This can be useful for
+   * debugging and understanding the model's reasoning process. One of `auto`,
+   * `concise`, or `detailed`.
+   */
+  summary?: 'auto' | 'concise' | 'detailed' | null;
 }
 
 /**
@@ -265,3 +289,11 @@ export interface ResponseFormatText {
    */
   type: 'text';
 }
+
+export type ResponsesModel =
+  | (string & {})
+  | ChatModel
+  | 'o1-pro'
+  | 'o1-pro-2025-03-19'
+  | 'computer-use-preview'
+  | 'computer-use-preview-2025-03-11';