+++ tests/api-resources/audio/speech.test.ts
@@ -13,7 +13,8 @@ describe('resource speech', () => {
     const response = await client.audio.speech.create({
       input: 'input',
       model: 'string',
-      voice: 'alloy',
+      voice: 'ash',
+      instructions: 'instructions',
       response_format: 'mp3',
       speed: 0.25,
     });